Refactor: Let's not make events its own app after all.
Will need to keep the app stub until deployed, since it's
responsible for tearing down the old tables.

This also triggered the old game/rules -> rules/accounts -> accounts/game
circular dependency chain. Let's try collapsing that by moving
rules back into game and making some offending foreign key references
strings.
