+++ camp/game/admin.py
@@ -28,3 +28,18 @@ class ChapterAdmin(admin.ModelAdmin):
 @admin.register(models.Campaign)
 class CampaignAdmin(admin.ModelAdmin):
     pass
+
+
+class RegistrationInline(admin.TabularInline):
+    model = models.EventRegistration
+
+
+@admin.register(models.Event)
+class EventAdmin(admin.ModelAdmin):
+    date_hierarchy = "created_date"
+    inlines = [RegistrationInline]
+
+
+@admin.register(models.EventRegistration)
+class EventRegistrationAdmin(admin.ModelAdmin):
+    date_hierarchy = "registered_date"
