+++ camp/game/templates/events/event_list.html
@@ -15,7 +15,7 @@ Events
     <li>No events scheduled.</li>
     {% endfor %}
   </ul>
-  {% has_perm 'events.add_event' user chapter as can_create_event %}
+  {% has_perm 'game.add_event' user chapter as can_create_event %}
   {% if can_create_event %}
   <p><a href="{% url 'event-create' chapter.slug %}">Create Event in {{chapter.name}}...</a></p>
   {% endif %}
